chage struct separate HttpRequestAgent from WebAppInterface
