+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/Campaign-SDK.iml" filepath="$PROJECT_DIR$/Campaign-SDK.iml" />
       <module fileurl="file://$PROJECT_DIR$/CampaignSDK.iml" filepath="$PROJECT_DIR$/CampaignSDK.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
