+++ app/src/main/java/com/adviser/campaign/WebAppInterface.kt
@@ -1,8 +1,6 @@
 package com.adviser.campaign
 
 import android.webkit.JavascriptInterface
-import org.json.JSONArray
-import org.json.JSONObject
 
 /**
  * Created by Kairos on 2017. 5. 25..
@@ -10,34 +8,17 @@ import org.json.JSONObject
 
 class WebAppInterface {
 
-    val req: String?
-    var urls: Array<String>? = null
-    var cur = -1
+    var agent : HttpRequestAgent
 
-    constructor(req: String?) {
-        this.req = req
-        reqParser()
-    }
-
-    fun reqParser() {
-        val urls: Array<String>?
-
-        val images: JSONArray = JSONObject(req).getJSONArray("images")
-        urls = Array(images.length(), { _ -> "" })
-
-        for (i in 0..images.length() - 1) {
-            urls[i] = (images[i] as JSONObject).getString("url")
-        }
-
-        this.urls = urls
-        cur = 0
+    constructor(HttpRequestAgent: HttpRequestAgent) {
+        this.agent = HttpRequestAgent
     }
 
     @JavascriptInterface
     fun getImageUrl(): String? {
-        print(urls!![cur])
+        print(agent.urls!![agent.cur])
         // TODO add Exception process
-        return urls!![cur++]
+        return agent.urls!![agent.cur++]
     }
 
     @JavascriptInterface
